# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlexScale",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/api",
      "variables": {}
    }
  ],
  "paths": {
    "/auth/verifyToken":{
      "get": {
        "tags": [
          "Login"
        ],
        "summary": "verifyToken",
        "operationId": "verifyToken",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "TOKEN"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/timesheet?{userId}": {
      "get": {
        "tags": [
          "Timesheet"
        ],
        "summary": "getTimesheetByUserId",
        "operationId": "getTimesheetByUserId",
        "parameters": [
          {
            "name": "orgId",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63877f9def951c55786bf2a4"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63947bca32551d4d32d5f1ca"
            }
          },
          {
            "name": "contractId",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63947cf332551d4d32d5f2ae"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      }
    },
    "/timesheet?{list}": {
      "get": {
        "tags": [
          "Timesheet"
        ],
        "summary": "getAllTimesheets",
        "operationId": "getAllTimesheets",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      }
    },
    "/timesheet": {
      "get": {
        "tags": [
          "Timesheet"
        ],
        "summary": "getTimesheetById",
        "operationId": "getTimesheetById",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63b2ba5821a798ee7ac10100"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      },
      "post": {
        "tags": [
          "Timesheet"
        ],
        "summary": "create",
        "operationId": "Postcreate1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createTimesheet"
                  },
                  {
                    "example": {
                      "CreatedBy": "6386106b8534e766fcbada14",
                      "TimePeriodStart": "12-26-2022",
                      "TimePeriodEnd": "2022-12-29",
                      "Organization": {
                        "_id": "63877f9def951c55786bf2a4",
                        "Name": "NCT",
                        "OrganizationType": "Client",
                        "CreatedBy": "6386106b8534e766fcbada14",
                        "Address1": "test address",
                        "Address2": "test address2",
                        "City": "Islamabad",
                        "State": "Punjab",
                        "Country": "Iraq",
                        "PostalCode": "3365",
                        "BillingContactEmail": "test@test.com",
                        "BillingContactPhone": "858585858",
                        "BillingContactFirstName": "Test",
                        "BillingContactLastName": "Test LName",
                        "BillingAddress": "test address",
                        "BillingAddress2": "test address2",
                        "BillingCity": "Islamabad",
                        "BillingState": "Punjab",
                        "BillingPostalCode": "3364",
                        "BillingCountry": "Russia",
                        "createdAt": "2022-11-30T16:06:53.107Z",
                        "updatedAt": "2022-12-29T09:45:51.638Z",
                        "__v": 0,
                        "StripeId": "cus_N3jeQGPwDZKyaf"
                      },
                      "User": {
                        "_id": "63947bca32551d4d32d5f1ca",
                        "FirstName": "John",
                        "PreferredName": "John",
                        "LastName": "Deo",
                        "Email": "john@deo.com",
                        "Phone": "12312312",
                        "AccessLevel": 6,
                        "Status": "Active",
                        "CreatedBy": "6386106b8534e766fcbada14",
                        "Location": "Algeria",
                        "Organization": [],
                        "Manager": [
                          {
                            "Documents": [],
                            "_id": "63877d43ef951c55786bf211",
                            "FirstName": "Muhammad",
                            "PreferredName": "Ali",
                            "LastName": "Ali",
                            "Email": "syedalisma1234@gmail.com",
                            "Phone": "03120116333",
                            "AccessLevel": 5,
                            "Status": "Active",
                            "CreatedBy": "6386106b8534e766fcbada14",
                            "Location": "Pakistan",
                            "Organization": [
                              "638610018534e766fcbada10",
                              "63877f9def951c55786bf2a4"
                            ],
                            "Manager": [],
                            "Password": "$2b$10$mvKqPtUPvX/xhdjJ/DZvSuiYlgiPI0rqeMstWeo8YFad3osune9FC",
                            "Gender": "He/Him",
                            "Contracts": [],
                            "createdAt": "2022-11-30T15:56:51.945Z",
                            "updatedAt": "2022-12-08T16:23:45.709Z",
                            "__v": 0
                          }
                        ],
                        "Password": "$2b$10$QvOJeRq3bM1DaTHS.wsNrubOCuhDk8aIaMgeKmjIMsx56VN0KC0Re",
                        "Gender": "He/Him",
                        "Contracts": [],
                        "Documents": [],
                        "createdAt": "2022-12-10T12:30:02.729Z",
                        "updatedAt": "2022-12-27T23:14:12.385Z",
                        "__v": 0
                      },
                      "Contract": {
                        "_id": "63947cf332551d4d32d5f2ae",
                        "JobTitle": "Data Engineer",
                        "CreatedBy": "63878963ef951c55786bf365",
                        "Status": "Onboarding",
                        "StartDate": "2022-12-09T23:00:00.000Z",
                        "Organization": "63877f9def951c55786bf2a4",
                        "TypeofEmployment": "None",
                        "User": "63947bca32551d4d32d5f1ca",
                        "Placement": "63947c1632551d4d32d5f21e",
                        "createdAt": "2022-12-10T12:34:59.676Z",
                        "updatedAt": "2022-12-27T12:17:37.648Z",
                        "__v": 0,
                        "EndDate": "2023-01-30T23:00:00.000Z",
                        "GoLiveDate": "2022-12-27T23:00:00.000Z",
                        "NoOfPaidTimeOffDays": 30,
                        "RatePerHour": 30,
                        "WorkHoursPerWeek": 20
                      },
                      "TimeLog": [
                        {
                          "date": "12-26-2022",
                          "hours": "8",
                          "check": "false"
                        },
                        {
                          "date": "2022-12-26T23:00:00.000Z",
                          "hours": "8",
                          "check": "false"
                        },
                        {
                          "date": "2022-12-27T23:00:00.000Z",
                          "hours": "0",
                          "check": "true"
                        },
                        {
                          "date": "2022-12-28T23:00:00.000Z",
                          "hours": "8",
                          "check": "false"
                        },
                        {
                          "date": "2022-12-29T23:00:00.000Z",
                          "hours": "8",
                          "check": "false"
                        }
                      ]
                    }
                  }
                ]
              },
              "example": {
                "CreatedBy": "6386106b8534e766fcbada14",
                "TimePeriodStart": "12-26-2022",
                "TimePeriodEnd": "2022-12-29",
                "Organization": {
                  "_id": "63877f9def951c55786bf2a4",
                  "Name": "NCT",
                  "OrganizationType": "Client",
                  "CreatedBy": "6386106b8534e766fcbada14",
                  "Address1": "test address",
                  "Address2": "test address2",
                  "City": "Islamabad",
                  "State": "Punjab",
                  "Country": "Iraq",
                  "PostalCode": "3365",
                  "BillingContactEmail": "test@test.com",
                  "BillingContactPhone": "858585858",
                  "BillingContactFirstName": "Test",
                  "BillingContactLastName": "Test LName",
                  "BillingAddress": "test address",
                  "BillingAddress2": "test address2",
                  "BillingCity": "Islamabad",
                  "BillingState": "Punjab",
                  "BillingPostalCode": "3364",
                  "BillingCountry": "Russia",
                  "createdAt": "2022-11-30T16:06:53.107Z",
                  "updatedAt": "2022-12-29T09:45:51.638Z",
                  "__v": 0,
                  "StripeId": "cus_N3jeQGPwDZKyaf"
                },
                "User": {
                  "_id": "63947bca32551d4d32d5f1ca",
                  "FirstName": "John",
                  "PreferredName": "John",
                  "LastName": "Deo",
                  "Email": "john@deo.com",
                  "Phone": "12312312",
                  "AccessLevel": 6,
                  "Status": "Active",
                  "CreatedBy": "6386106b8534e766fcbada14",
                  "Location": "Algeria",
                  "Organization": [],
                  "Manager": [
                    {
                      "Documents": [],
                      "_id": "63877d43ef951c55786bf211",
                      "FirstName": "Muhammad",
                      "PreferredName": "Ali",
                      "LastName": "Ali",
                      "Email": "syedalisma1234@gmail.com",
                      "Phone": "03120116333",
                      "AccessLevel": 5,
                      "Status": "Active",
                      "CreatedBy": "6386106b8534e766fcbada14",
                      "Location": "Pakistan",
                      "Organization": [
                        "638610018534e766fcbada10",
                        "63877f9def951c55786bf2a4"
                      ],
                      "Manager": [],
                      "Password": "$2b$10$mvKqPtUPvX/xhdjJ/DZvSuiYlgiPI0rqeMstWeo8YFad3osune9FC",
                      "Gender": "He/Him",
                      "Contracts": [],
                      "createdAt": "2022-11-30T15:56:51.945Z",
                      "updatedAt": "2022-12-08T16:23:45.709Z",
                      "__v": 0
                    }
                  ],
                  "Password": "$2b$10$QvOJeRq3bM1DaTHS.wsNrubOCuhDk8aIaMgeKmjIMsx56VN0KC0Re",
                  "Gender": "He/Him",
                  "Contracts": [],
                  "Documents": [],
                  "createdAt": "2022-12-10T12:30:02.729Z",
                  "updatedAt": "2022-12-27T23:14:12.385Z",
                  "__v": 0
                },
                "Contract": {
                  "_id": "63947cf332551d4d32d5f2ae",
                  "JobTitle": "Data Engineer",
                  "CreatedBy": "63878963ef951c55786bf365",
                  "Status": "Onboarding",
                  "StartDate": "2022-12-09T23:00:00.000Z",
                  "Organization": "63877f9def951c55786bf2a4",
                  "TypeofEmployment": "None",
                  "User": "63947bca32551d4d32d5f1ca",
                  "Placement": "63947c1632551d4d32d5f21e",
                  "createdAt": "2022-12-10T12:34:59.676Z",
                  "updatedAt": "2022-12-27T12:17:37.648Z",
                  "__v": 0,
                  "EndDate": "2023-01-30T23:00:00.000Z",
                  "GoLiveDate": "2022-12-27T23:00:00.000Z",
                  "NoOfPaidTimeOffDays": 30,
                  "RatePerHour": 30,
                  "WorkHoursPerWeek": 20
                },
                "TimeLog": [
                  {
                    "date": "12-26-2022",
                    "hours": "8",
                    "check": "false"
                  },
                  {
                    "date": "2022-12-26T23:00:00.000Z",
                    "hours": "8",
                    "check": "false"
                  },
                  {
                    "date": "2022-12-27T23:00:00.000Z",
                    "hours": "0",
                    "check": "true"
                  },
                  {
                    "date": "2022-12-28T23:00:00.000Z",
                    "hours": "8",
                    "check": "false"
                  },
                  {
                    "date": "2022-12-29T23:00:00.000Z",
                    "hours": "8",
                    "check": "false"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      },
      "put": {
        "tags": [
          "Timesheet"
        ],
        "summary": "updateTimesheet",
        "operationId": "updateTimesheet",
        "parameters": [
          {
            "name": "timesheetId",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63a58e98e7f5ae01f1199ccd"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/updateTimesheetRequest"
                  },
                  {
                    "example": {
                      "TimePeriodStart": "12-26-2022",
                      "TimePeriodEnd": "2022-12-29",
                      "TimeLog": [
                        {
                          "date": "12-26-2022",
                          "hours": "8",
                          "check": "false"
                        },
                        {
                          "date": "2022-12-26T23:00:00.000Z",
                          "hours": "8",
                          "check": "false"
                        },
                        {
                          "date": "2022-12-27T23:00:00.000Z",
                          "hours": "0",
                          "check": "true"
                        },
                        {
                          "date": "2022-12-28T23:00:00.000Z",
                          "hours": "8",
                          "check": "false"
                        },
                        {
                          "date": "2022-12-29T23:00:00.000Z",
                          "hours": "8",
                          "check": "false"
                        }
                      ]
                    }
                  }
                ]
              },
              "example": {
                "TimePeriodStart": "12-26-2022",
                "TimePeriodEnd": "2022-12-29",
                "TimeLog": [
                  {
                    "date": "12-26-2022",
                    "hours": "8",
                    "check": "false"
                  },
                  {
                    "date": "2022-12-26T23:00:00.000Z",
                    "hours": "8",
                    "check": "false"
                  },
                  {
                    "date": "2022-12-27T23:00:00.000Z",
                    "hours": "0",
                    "check": "true"
                  },
                  {
                    "date": "2022-12-28T23:00:00.000Z",
                    "hours": "8",
                    "check": "false"
                  },
                  {
                    "date": "2022-12-29T23:00:00.000Z",
                    "hours": "8",
                    "check": "false"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      },
      "delete": {
        "tags": [
          "Timesheet"
        ],
        "summary": "deleteTimesheet",
        "operationId": "deleteTimesheet",
        "parameters": [
          {
            "name": "timesheetId",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63a58e98e7f5ae01f1199ccd"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
    },
    "schemas": {
...
      "Organization": {
        "title": "Organization",
        "required": [
          "_id",
          "Name",
          "OrganizationType",
          "CreatedBy",
          "Address1",
          "Address2",
          "City",
          "State",
          "Country",
          "PostalCode",
          "BillingContactEmail",
          "BillingContactPhone",
          "BillingContactFirstName",
          "BillingContactLastName",
          "BillingAddress",
          "BillingAddress2",
          "BillingCity",
          "BillingState",
          "BillingPostalCode",
          "BillingCountry",
          "createdAt",
          "updatedAt",
          "__v",
          "StripeId"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "OrganizationType": {
            "type": "string"
          },
          "CreatedBy": {
            "type": "string"
          },
          "Address1": {
            "type": "string"
          },
          "Address2": {
            "type": "string"
          },
          "City": {
            "type": "string"
          },
          "State": {
            "type": "string"
          },
          "Country": {
            "type": "string"
          },
          "PostalCode": {
            "type": "string"
          },
          "BillingContactEmail": {
            "type": "string"
          },
          "BillingContactPhone": {
            "type": "string"
          },
          "BillingContactFirstName": {
            "type": "string"
          },
          "BillingContactLastName": {
            "type": "string"
          },
          "BillingAddress": {
            "type": "string"
          },
          "BillingAddress2": {
            "type": "string"
          },
          "BillingCity": {
            "type": "string"
          },
          "BillingState": {
            "type": "string"
          },
          "BillingPostalCode": {
            "type": "string"
          },
          "BillingCountry": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "__v": {
            "type": "integer",
            "format": "int32"
          },
          "StripeId": {
            "type": "string"
          }
        },
        "example": {
          "_id": "63877f9def951c55786bf2a4",
          "Name": "NCT",
          "OrganizationType": "Client",
          "CreatedBy": "6386106b8534e766fcbada14",
          "Address1": "test address",
          "Address2": "test address2",
          "City": "Islamabad",
          "State": "Punjab",
          "Country": "Iraq",
          "PostalCode": "3365",
          "BillingContactEmail": "test@test.com",
          "BillingContactPhone": "858585858",
          "BillingContactFirstName": "Test",
          "BillingContactLastName": "Test LName",
          "BillingAddress": "test address",
          "BillingAddress2": "test address2",
          "BillingCity": "Islamabad",
          "BillingState": "Punjab",
          "BillingPostalCode": "3364",
          "BillingCountry": "Russia",
          "createdAt": "2022-11-30T16:06:53.107Z",
          "updatedAt": "2022-12-29T09:45:51.638Z",
          "__v": 0,
          "StripeId": "cus_N3jeQGPwDZKyaf"
        }
      },
...
      "createTimesheet": {
        "title": "createTimesheet",
        "required": [
          "CreatedBy",
          "TimePeriodStart",
          "TimePeriodEnd",
          "Organization",
          "User",
          "Contract",
          "TimeLog"
        ],
        "type": "object",
        "properties": {
          "CreatedBy": {
            "type": "string"
          },
          "TimePeriodStart": {
            "type": "string"
          },
          "TimePeriodEnd": {
            "type": "string"
          },
          "Organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "User": {
            "$ref": "#/components/schemas/User"
          },
          "Contract": {
            "$ref": "#/components/schemas/Contract1"
          },
          "TimeLog": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeLog"
            },
            "description": ""
          }
        },
        "example": {
          "CreatedBy": "6386106b8534e766fcbada14",
          "TimePeriodStart": "12-26-2022",
          "TimePeriodEnd": "2022-12-29",
          "Organization": {
            "_id": "63877f9def951c55786bf2a4",
            "Name": "NCT",
            "OrganizationType": "Client",
            "CreatedBy": "6386106b8534e766fcbada14",
            "Address1": "test address",
            "Address2": "test address2",
            "City": "Islamabad",
            "State": "Punjab",
            "Country": "Iraq",
            "PostalCode": "3365",
            "BillingContactEmail": "test@test.com",
            "BillingContactPhone": "858585858",
            "BillingContactFirstName": "Test",
            "BillingContactLastName": "Test LName",
            "BillingAddress": "test address",
            "BillingAddress2": "test address2",
            "BillingCity": "Islamabad",
            "BillingState": "Punjab",
            "BillingPostalCode": "3364",
            "BillingCountry": "Russia",
            "createdAt": "2022-11-30T16:06:53.107Z",
            "updatedAt": "2022-12-29T09:45:51.638Z",
            "__v": 0,
            "StripeId": "cus_N3jeQGPwDZKyaf"
          },
          "User": {
            "_id": "63947bca32551d4d32d5f1ca",
            "FirstName": "John",
            "PreferredName": "John",
            "LastName": "Deo",
            "Email": "john@deo.com",
            "Phone": "12312312",
            "AccessLevel": 6,
            "Status": "Active",
            "CreatedBy": "6386106b8534e766fcbada14",
            "Location": "Algeria",
            "Organization": [],
            "Manager": [
              {
                "Documents": [],
                "_id": "63877d43ef951c55786bf211",
                "FirstName": "Muhammad",
                "PreferredName": "Ali",
                "LastName": "Ali",
                "Email": "syedalisma1234@gmail.com",
                "Phone": "03120116333",
                "AccessLevel": 5,
                "Status": "Active",
                "CreatedBy": "6386106b8534e766fcbada14",
                "Location": "Pakistan",
                "Organization": [
                  "638610018534e766fcbada10",
                  "63877f9def951c55786bf2a4"
                ],
                "Manager": [],
                "Password": "$2b$10$mvKqPtUPvX/xhdjJ/DZvSuiYlgiPI0rqeMstWeo8YFad3osune9FC",
                "Gender": "He/Him",
                "Contracts": [],
                "createdAt": "2022-11-30T15:56:51.945Z",
                "updatedAt": "2022-12-08T16:23:45.709Z",
                "__v": 0
              }
            ],
            "Password": "$2b$10$QvOJeRq3bM1DaTHS.wsNrubOCuhDk8aIaMgeKmjIMsx56VN0KC0Re",
            "Gender": "He/Him",
            "Contracts": [],
            "Documents": [],
            "createdAt": "2022-12-10T12:30:02.729Z",
            "updatedAt": "2022-12-27T23:14:12.385Z",
            "__v": 0
          },
          "Contract": {
            "_id": "63947cf332551d4d32d5f2ae",
            "JobTitle": "Data Engineer",
            "CreatedBy": "63878963ef951c55786bf365",
            "Status": "Onboarding",
            "StartDate": "2022-12-09T23:00:00.000Z",
            "Organization": "63877f9def951c55786bf2a4",
            "TypeofEmployment": "None",
            "User": "63947bca32551d4d32d5f1ca",
            "Placement": "63947c1632551d4d32d5f21e",
            "createdAt": "2022-12-10T12:34:59.676Z",
            "updatedAt": "2022-12-27T12:17:37.648Z",
            "__v": 0,
            "EndDate": "2023-01-30T23:00:00.000Z",
            "GoLiveDate": "2022-12-27T23:00:00.000Z",
            "NoOfPaidTimeOffDays": 30,
            "RatePerHour": 30,
            "WorkHoursPerWeek": 20
          },
          "TimeLog": [
            {
              "date": "12-26-2022",
              "hours": "8",
              "check": "false"
            },
            {
              "date": "2022-12-26T23:00:00.000Z",
              "hours": "8",
              "check": "false"
            },
            {
              "date": "2022-12-27T23:00:00.000Z",
              "hours": "0",
              "check": "true"
            },
            {
              "date": "2022-12-28T23:00:00.000Z",
              "hours": "8",
              "check": "false"
            },
            {
              "date": "2022-12-29T23:00:00.000Z",
              "hours": "8",
              "check": "false"
            }
          ]
        }
      },
...
      "TimeLog": {
        "title": "TimeLog",
        "required": [
          "date",
          "hours",
          "check"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "hours": {
            "type": "string"
          },
          "check": {
            "type": "string"
          }
        },
        "example": {
          "date": "12-26-2022",
          "hours": "8",
          "check": "false"
        }
      },
      "updateTimesheetRequest": {
        "title": "updateTimesheetRequest",
        "required": [
          "TimePeriodStart",
          "TimePeriodEnd",
          "TimeLog"
        ],
        "type": "object",
        "properties": {
          "TimePeriodStart": {
            "type": "string"
          },
          "TimePeriodEnd": {
            "type": "string"
          },
          "TimeLog": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeLog"
            },
            "description": ""
          }
        },
        "example": {
          "TimePeriodStart": "12-26-2022",
          "TimePeriodEnd": "2022-12-29",
          "TimeLog": [
            {
              "date": "12-26-2022",
              "hours": "8",
              "check": "false"
            },
            {
              "date": "2022-12-26T23:00:00.000Z",
              "hours": "8",
              "check": "false"
            },
            {
              "date": "2022-12-27T23:00:00.000Z",
              "hours": "0",
              "check": "true"
            },
            {
              "date": "2022-12-28T23:00:00.000Z",
              "hours": "8",
              "check": "false"
            },
            {
              "date": "2022-12-29T23:00:00.000Z",
              "hours": "8",
              "check": "false"
            }
          ]
        }
      },
...
    }
  },
...
}